% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fairness.R
\name{check_cond_statistical_parity}
\alias{check_cond_statistical_parity}
\title{Examine conditional statistical parity of a model}
\usage{
check_cond_statistical_parity(
  data,
  outcome,
  group,
  group2,
  condition,
  probs,
  cutoff = 0.5,
  confint = TRUE,
  bootstraps = 1000,
  message = TRUE,
  digits = 2
)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
sensitive attribute}

\item{outcome}{Name of the outcome variable, it must be binary}

\item{group}{Name of the sensitive attribute}

\item{group2}{Name of the group to condition on}

\item{condition}{If the conditional group is categorical, the condition
supplied must be a character of the levels to condition on. If the conditional
group is continuous, the conditions supplied must be a character containing
the sign of the condition and the value to threshold the continuous variable
(e.g. "<50", ">50", "<=50", ">=50").}

\item{probs}{Name of the predicted outcome variable}

\item{cutoff}{Threshold for the predicted outcome, default is 0.5}

\item{confint}{Whether to compute 95\% confidence interval, default is TRUE}

\item{bootstraps}{Number of bootstrap samples, default is 1000}

\item{message}{Whether to print the results, default is TRUE}

\item{digits}{Number of digits to round the results to, default is 2}
}
\value{
A list containing the following elements:
\itemize{
\item Conditions: The conditions used to calculate the conditional PPR
\item PPR_Group1: Positive Prediction Rate for the first group
\item PPR_Group2: Positive Prediction Rate for the second group
\item PPR_Diff: Difference in Positive Prediction Rate
If confidence intervals are computed (\code{confint = TRUE}):
\item PPR_Diff_CI: 95\% confidence interval for the difference in Positive
Prediction Rate
}
}
\description{
Examine conditional statistical parity of a model
}
