% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fairness.R
\name{conditional_statistical_parity}
\alias{conditional_statistical_parity}
\title{Examine conditional statistical parity of a model}
\usage{
conditional_statistical_parity(
  data,
  outcome,
  group,
  group2,
  probs,
  cutoff = 0.5,
  group2_cutoff,
  confint = TRUE,
  bootstraps = 1000
)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
sensitive attributes}

\item{outcome}{the name of the outcome variable, it must be binary}

\item{group}{the name of the sensitive attribute}

\item{group2}{the name of the conditional sensitive attribute}

\item{probs}{the name of the predicted outcome variable}

\item{cutoff}{the threshold for the predicted outcome, default is 0.5}

\item{group2_cutoff}{the threshold for the conditional sensitive attribute.}

\item{confint}{whether to compute 95\% confidence interval, default is TRUE}

\item{bootstraps}{the number of bootstrap samples, default is 1000}
}
\value{
a list of conditional statistical parity and its confidence interval
}
\description{
Examine conditional statistical parity of a model
}
