% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fairness.R
\name{eval_bs_parity}
\alias{eval_bs_parity}
\title{Examine Brier Score parity of a model}
\usage{
eval_bs_parity(
  data,
  outcome,
  group,
  probs,
  alpha = 0.05,
  bootstraps = 2500,
  digits = 2,
  message = TRUE
)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
sensitive attribute}

\item{outcome}{Name of the outcome variable}

\item{group}{Name of the sensitive attribute}

\item{probs}{Predicted probabilities}

\item{alpha}{The 1 - significance level for the confidence interval, default is 0.05}

\item{bootstraps}{Number of bootstraps to use for confidence intervals}

\item{digits}{Number of digits to round the results to, default is 2}

\item{message}{Whether to print the results, default is TRUE}

\item{confint}{Logical indicating whether to calculate confidence intervals}
}
\value{
A list containing the following elements:
\itemize{
\item Brier Score for Group 1
\item Brier Score for Group 2
\item Difference in Brier Score
If confidence intervals are computed (\code{confint = TRUE}):
\item A vector of length 2 containing the lower and upper bounds of the 95\%
confidence interval for the difference in Brier Score
}
}
\description{
Examine Brier Score parity of a model
}
