% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGroupFairness.R
\name{eval_generalized_entropy_index}
\alias{eval_generalized_entropy_index}
\title{Examine Generalized Entropy Index of a model}
\usage{
eval_generalized_entropy_index(
  data,
  outcome,
  group,
  probs,
  get_metric,
  alpha = 2,
  cutoff = 0.5,
  confint = TRUE,
  bootstraps = 1000,
  digits = 2,
  message = TRUE
)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
group}

\item{outcome}{Name of the outcome variable, it must be binary}

\item{group}{Name of the group}

\item{probs}{Name of the predicted outcome variable}

\item{get_metric}{Function of the model performance metric used to compute multi-group metrics
The input should take in: data, outcome, group, probs, cutoff, and digits
The output should be a list containing the following elements:
\itemize{
\item Name of each group
\item Calculated metric for each group
}}

\item{alpha}{Alpha for the generalized entropy index function (Cannot be 0 or 1), default is 2}

\item{cutoff}{Threshold for the predicted outcome, default is 0.5}

\item{confint}{Whether to compute 95\% confidence interval, default is TRUE}

\item{bootstraps}{Number of bootstrap samples, default is 1000}

\item{digits}{Number of digits to round the results to, default is 2}

\item{message}{Whether to print the results, default is TRUE}
}
\value{
A list containing the following elements:
\itemize{
\item generalized_entropy_index: generalized entropy index for the model
If confidence intervals are computed (\code{confint = TRUE}):
\item GEN_ENTROPY_IND_CI: A vector of length 2 providing the lower and upper bounds of
the 95\% confidence interval for the generalized entropy index
}
}
\description{
Examine Generalized Entropy Index of a model
}
