% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGroupFairness.R
\name{eval_mean_abs_dev}
\alias{eval_mean_abs_dev}
\title{Examine mean absolute deviation of a model
#' This function evaluates the mean absolute deviation in model
performance metrics across different groups.}
\usage{
eval_mean_abs_dev(data, outcome, group, probs, cutoff = 0.5, digits = 2)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
group}

\item{outcome}{Name of the outcome variable, it must be binary}

\item{group}{Name of the group}

\item{probs}{Name of the predicted outcome variable}

\item{cutoff}{Threshold for the predicted outcome, default is 0.5}

\item{digits}{Number of digits to round the results to, default is 2}
}
\value{
A data frame containing the following elements:
\itemize{
\item Metric: The names of the metrics calculated.
\item Values for each group.
\item Mean_abs_dev: The mean absolute deviation for the model
}
}
\description{
Examine mean absolute deviation of a model
#' This function evaluates the mean absolute deviation in model
performance metrics across different groups.
}
