% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fairness.R
\name{eval_neg_class_bal}
\alias{eval_neg_class_bal}
\title{Examine balance for negative class of a model}
\usage{
eval_neg_class_bal(
  data,
  outcome,
  group,
  probs,
  confint = TRUE,
  bootstraps = 2500,
  digits = 2,
  message = TRUE
)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
sensitive attribute}

\item{outcome}{Name of the outcome variable}

\item{group}{Name of the sensitive attribute}

\item{probs}{Predicted probabilities}

\item{confint}{Logical indicating whether to calculate confidence intervals}

\item{bootstraps}{Number of bootstraps to use for confidence intervals}

\item{digits}{Number of digits to round the results to, default is 2}

\item{message}{Whether to print the results, default is TRUE}
}
\value{
A list containing the following elements:
\itemize{
\item Average predicted probability for Group 1
\item Average predicted probability for Group 2
\item Difference in average predicted probability
If confidence intervals are computed (\code{confint = TRUE}):
\item A vector of length 2 containing the lower and upper bounds of the 95\%
confidence interval for the difference in average predicted probability
}
}
\description{
Examine balance for negative class of a model
}
