% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fairness.R
\name{eval_pred_parity}
\alias{eval_pred_parity}
\title{Examine predictive parity of a model}
\usage{
eval_pred_parity(
  data,
  outcome,
  group,
  probs,
  cutoff = 0.5,
  confint = TRUE,
  bootstraps = 2500,
  digits = 2,
  message = TRUE
)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
sensitive attribute}

\item{outcome}{Name of the outcome variable, it must be binary}

\item{group}{Name of the sensitive attribute}

\item{probs}{Name of the predicted outcome variable}

\item{cutoff}{Threshold for the predicted outcome, default is 0.5}

\item{confint}{Whether to compute 95\% confidence interval, default is TRUE}

\item{bootstraps}{Number of bootstrap samples, default is 2500}

\item{digits}{Number of digits to round the results to, default is 2}

\item{message}{Whether to print the results, default is TRUE}
}
\value{
A list containing the following elements:
\itemize{
\item PPV_Group1: Positive Predictive Value for the first group
\item PPV_Group2: Positive Predictive Value for the second group
\item PPV_Diff: Difference in Positive Predictive Value
If confidence intervals are computed (\code{confint = TRUE}):
\item PPV_Diff_CI: A vector of length 2 containing the lower and upper bounds
of the 95\% confidence interval for the difference in Positive Predictive
Value
}
}
\description{
Examine predictive parity of a model
}
