% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fairness.R
\name{eval_stats_parity}
\alias{eval_stats_parity}
\title{Examine statistical parity of a model}
\usage{
eval_stats_parity(
  data,
  outcome,
  group,
  probs,
  cutoff = 0.5,
  confint = TRUE,
  bootstraps = 1000,
  digits = 2,
  message = TRUE
)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
sensitive attribute}

\item{outcome}{Name of the outcome variable, it must be binary}

\item{group}{Name of the sensitive attribute}

\item{probs}{Name of the predicted outcome variable}

\item{cutoff}{Threshold for the predicted outcome, default is 0.5}

\item{confint}{Whether to compute 95\% confidence interval, default is TRUE}

\item{bootstraps}{Number of bootstrap samples, default is 1000}

\item{digits}{Number of digits to round the results to, default is 2}

\item{message}{Whether to print the results, default is TRUE}
}
\value{
A list containing the following elements:
\itemize{
\item PPR_Group1: Positive Prediction Rate for the first group
\item PPR_Group2: Positive Prediction Rate for the second group
\item PPR_Diff: Difference in Positive Prediction Rate
If confidence intervals are computed (\code{confint = TRUE}):
\item PPR_Diff_CI: 95\% confidence interval for the difference in Positive
Prediction Rate
}
}
\description{
Examine statistical parity of a model
}
