% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{fit_rf}
\alias{fit_rf}
\title{Fit a random forest model}
\usage{
fit_rf(data)
}
\arguments{
\item{data}{a dataframe object}
}
\value{
Dataframe with outcome, predicted probability, and protected
attributes
}
\description{
Fit a random forest model
}
