% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPerformance.R
\name{get_all_metrics}
\alias{get_all_metrics}
\title{Calculate the all metrics at once}
\usage{
get_all_metrics(data, outcome, group, probs, cutoff = 0.5, digits = 2)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
sensitive attribute}

\item{outcome}{the name of the outcome variable, it must be binary}

\item{group}{the name of the sensitive attribute}

\item{probs}{the name of the predicted outcome variable}

\item{cutoff}{the threshold for the predicted outcome, default is 0.5}

\item{digits}{the number of digits to round the result to, default is 2}
}
\value{
a Data frame of all metrics
}
\description{
Calculate the all metrics at once
}
