% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fairness.R
\name{predictive_parity}
\alias{predictive_parity}
\title{Examine predictive parity of a model}
\usage{
predictive_parity(
  data,
  outcome,
  group,
  probs,
  cutoff = 0.5,
  confint = TRUE,
  bootstraps = 1000
)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
sensitive attribute}

\item{outcome}{the name of the outcome variable, it must be binary}

\item{group}{the name of the sensitive attribute}

\item{probs}{the name of the predicted outcome variable}

\item{cutoff}{the threshold for the predicted outcome, default is 0.5}

\item{confint}{whether to compute 95\% confidence interval, default is TRUE}

\item{bootstraps}{the number of bootstrap samples, default is 1000}
}
\value{
a list of predictive parity and its confidence interval
}
\description{
Examine predictive parity of a model
}
