% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fairness.R
\name{eval_eq_odds}
\alias{eval_eq_odds}
\title{Examine equalized odds of a model}
\usage{
eval_eq_odds(
  data,
  outcome,
  group,
  probs,
  cutoff = 0.5,
  confint = TRUE,
  bootstraps = 1000,
  digits = 2,
  message = TRUE
)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
sensitive attribute}

\item{outcome}{Name of the outcome variable, it must be binary}

\item{group}{Name of the sensitive attribute}

\item{probs}{Name of the predicted outcome variable}

\item{cutoff}{Threshold for the predicted outcome, default is 0.5}

\item{confint}{Whether to compute 95\% confidence interval, default is TRUE}

\item{bootstraps}{Number of bootstrap samples, default is 1000}

\item{digits}{Number of digits to round the results to, default is 2}

\item{message}{Whether to print the results, default is TRUE}
}
\value{
A list containing the following elements:
\itemize{
\item TPR_Group1: True Positive Rate for the first group
\item TPR_Group2: True Positive Rate for the second group
\item TPR_Difference: The difference in True Positive Rates between the two
groups
\item FPR_Group1: False Positive Rate for the first group
\item FPR_Group2: False Positive Rate for the second group
\item FPR_Difference: The difference in False Positive Rates between the two
groups
If confidence intervals are computed (\code{confint = TRUE}):
\item TPR_Diff_CI: A vector of length 2 providing the lower and upper bounds
of the 95\% confidence interval for the TPR difference
\item FPR_Diff_CI: A vector of length 2 providing the lower and upper bounds
of the 95\% confidence interval for the FPR difference
}
}
\description{
Examine equalized odds of a model
}
