% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGroupFairness.R
\name{eval_max_min_diff}
\alias{eval_max_min_diff}
\title{Examine maximum minimum difference of a model}
\usage{
eval_max_min_diff(
  data,
  outcome,
  group,
  probs,
  get_metric,
  cutoff = 0.5,
  confint = TRUE,
  bootstraps = 1000,
  digits = 2,
  message = TRUE
)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
group}

\item{outcome}{Name of the outcome variable, it must be binary}

\item{group}{Name of the group}

\item{probs}{Name of the predicted outcome variable}

\item{get_metric}{Function of the model performance metric used to compute multi-group metrics
The input should take in: data, outcome, group, probs, cutoff, and digits
The output should be a list containing the following elements:
\itemize{
\item Name of each group
\item Calculated metric for each group
}}

\item{cutoff}{Threshold for the predicted outcome, default is 0.5}

\item{confint}{Whether to compute 95\% confidence interval, default is TRUE}

\item{bootstraps}{Number of bootstrap samples, default is 1000}

\item{digits}{Number of digits to round the results to, default is 2}

\item{message}{Whether to print the results, default is TRUE}
}
\value{
A list containing the following elements:
\itemize{
\item max_min_diff: maximum minimum metric difference for the model
If confidence intervals are computed (\code{confint = TRUE}):
\item MAX_MIN_DIFF_CI: A vector of length 2 providing the lower and upper bounds of
the 95\% confidence interval for the max-min metric difference
}
}
\description{
Examine maximum minimum difference of a model
}
