% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGroupFairness.R
\name{eval_max_min_ratio}
\alias{eval_max_min_ratio}
\title{Examine Maximum Minimum Ratio of a model}
\usage{
eval_max_min_ratio(data, outcome, group, probs, cutoff = 0.5, digits = 2)
}
\arguments{
\item{data}{Data frame containing the outcome, predicted outcome, and
group}

\item{outcome}{Name of the outcome variable, it must be binary}

\item{group}{Name of the group}

\item{probs}{Name of the predicted outcome variable}

\item{cutoff}{Threshold for the predicted outcome, default is 0.5}

\item{digits}{Number of digits to round the results to, default is 2}
}
\value{
A data frame containing the following elements:
\itemize{
\item Metric: The names of the metrics calculated.
\item Values for each group.
\item Max/Min: The maximum over minimum metric difference for the model
}
}
\description{
This function evaluates the maximum and minimum ratio in model
performance metrics across different groups.
}
